normalize amounts

+++ defitrack-rest/defitrack-api/src/main/java/io/defitrack/borrowing/vo/BorrowElementVO.kt
@@ -3,7 +3,6 @@ package io.defitrack.borrowing.vo
 import io.defitrack.network.NetworkVO
 import io.defitrack.protocol.ProtocolVO
 import io.defitrack.token.FungibleToken
-import java.math.BigInteger
 
 data class BorrowElementVO(
     val id: String,
@@ -12,6 +11,6 @@ data class BorrowElementVO(
     val dollarValue: Double,
     val name: String,
     val rate: Double?,
-    val amount: BigInteger,
+    val amount: Double,
     val token: FungibleToken
 )
